style(d1-pac): bump svd2rust to v0.27.1

+++ d1-pac/src/dmac/dmac_desc_addr.rs
@@ -53,7 +53,7 @@ impl R {
     #[doc = "Bits 2:31 - Lower 30 bits of DMA channel descriptor address"]
     #[inline(always)]
     pub fn dma_desc_addr(&self) -> DMA_DESC_ADDR_R {
-        DMA_DESC_ADDR_R::new(((self.bits >> 2) & 0x3fff_ffff) as u32)
+        DMA_DESC_ADDR_R::new((self.bits >> 2) & 0x3fff_ffff)
     }
 }
 impl W {
